An empty line was added to visually divide groups of the properties and the methods
